requirements.txt
Flask>=3.0
Pillow>=9.0
opencv-python-headless>=4.8
pymupdf>=1.23
fastapi>=0.110
uvicorn[standard]>=0.28
requests>=2.31
llama-cpp-python>=0.2.90
